somewhat working, added validation, and improved password handling overall

+++ internal/core/service/task_crud.go
@@ -2,6 +2,7 @@ package service
 
 import (
 	"context"
+	"fmt"
 	"strconv"
 
 	"github.com/Emin-07/TaskManager/internal/core/domain"
@@ -11,6 +12,9 @@ func (t TaskServ) Get(ctx context.Context, id string) (*domain.Task, error) {
 	idInt, err := strconv.Atoi(id)
 	if err != nil {
 		return nil, err
+	}
+	if idInt < 1 {
+		return nil, fmt.Errorf("there is no task with negative or zero id")
 	}
 	task, err := t.repo.Get(ctx, idInt)
 	if err != nil {
@@ -69,6 +73,12 @@ func (t TaskServ) List(ctx context.Context, limit, offset string) ([]*domain.Tas
 
 }
 func (t TaskServ) Post(ctx context.Context, task *domain.CreateTask, userId int) error {
+	if task.Title == "" {
+		return fmt.Errorf("can't create task with title")
+	}
+	if userId < 1 {
+		return fmt.Errorf("there is no user with negative or zero id")
+	}
 	return t.repo.Insert(ctx, task.Title, task.Text, task.Priority, task.ExpireDays, userId)
 }
 
@@ -77,6 +87,9 @@ func (t TaskServ) Delete(ctx context.Context, id string) error {
 	if err != nil {
 		return err
 	}
+	if idInt < 1 {
+		return fmt.Errorf("there is no task with negative or zero id")
+	}
 	return t.repo.Delete(ctx, idInt)
 }
 
@@ -84,6 +97,9 @@ func (t TaskServ) Patch(ctx context.Context, task *domain.CreateTask, id string)
 	idInt, err := strconv.Atoi(id)
 	if err != nil {
 		return err
+	}
+	if idInt < 1 {
+		return fmt.Errorf("there is no task with negative or zero id")
 	}
 	return t.repo.Patch(ctx, task.Title, task.Text, task.Priority, task.ExpireDays, idInt)
 }
